fix, rename: 수정된 Input 컴포넌트와 맞게 수정, 컴포넌트 폴더 생성 후 파일 위치 변경(#34)
