+++ src/app/(auth)/signin/_components/SignInForm/SignInForm.module.scss
@@ -1,0 +1,46 @@
+@import "@/styles/_index";
+@import "@/styles/_media";
+
+.container,
+.inputList,
+.inputBox {
+  display: flex;
+  flex-direction: column;
+}
+
+.container {
+  gap: 30px;
+  width: 100%;
+
+  .inputList {
+    gap: 16px;
+
+    .inputBox {
+      gap: 8px;
+
+      .label {
+        @include text-normal;
+        @include font-normal;
+
+        @include respond-to(tablet) {
+          @include text-sm;
+        }
+      }
+    }
+
+    @include respond-to(mobile) {
+      gap: 8px;
+    }
+  }
+
+  .signInButton {
+    width: auto;
+    padding: 21px;
+    @include text-lg;
+
+    @include respond-to(tablet) {
+      @include text-normal;
+      padding: 18px;
+    }
+  }
+}
